Allow for a sort-of monkey patch API for start logging setup.
common.start.main() now takes a monkey_log flag param which
will cause the Logger created by the setup (which routes logs
to a file or stdout depending on startup params) to be monkey
patched into the logging.root object so that module level
calls will us it. So now code can just import logging and call
logging.info(), for example, and it will log correctly without
needing to have a handle on the Logger instance itself.